+++ src/activations.c
@@ -1,8 +1,8 @@
 #include <activations.h>
 #include <math.h>
 
-float sign(float x) { return ((x >= 0) ? 1 : -1); }
+double sign(double x) { return ((x >= 0) ? 1 : -1); }
 
-float sigmoid(float x) { return 1.0f / (1.0f + (float)exp(-x)); }
+double sigmoid(double x) { return 1.0f / (1.0f + (double)exp(-x)); }
 
-float identity(float x) { return x; }
+double identity(double x) { return x; }
