+++ src/mathutils.c
@@ -1,18 +1,18 @@
 #include <mathutils.h>
 
-float vec_dot(size_t n, float *u, float *v) {
-  float d = 0;
+double vec_dot(size_t n, double *u, double *v) {
+  double d = 0;
   for (size_t i = 0; i < n; i++)
     d += u[i] * v[i];
   return d;
 }
 
-void vec_scale(size_t n, float k, float *u) {
+void vec_scale(size_t n, double k, double *u) {
   for (size_t i = 0; i < n; i++)
     u[i] *= k;
 }
 
-void vec_add(size_t n, float *u, const float *v) {
+void vec_add(size_t n, double *u, const double *v) {
   for (size_t i = 0; i < n; i++)
     u[i] += v[i];
 }
